Move client adapter to its own package
