+++ internal/client/adapter/client_adapter_test.go
@@ -1,7 +1,7 @@
 // Copyright 2025 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
-package client
+package adapter
 
 import (
 	"testing"
